==> Attraction_Recommenation/step04_01-preprocessing.py <==
import pandas as pd
from konlpy.tag import Okt
import re

df = pd.read_csv('./crawling_data/merged_reviews_trip_naver.csv')
df.info()

okt = Okt()

stopwords = pd.read_csv('./crawling_data/stopwords.csv', index_col=0)
stopwords_add = ['리뷰']
stopwords = list(stopwords['stopword']) + stopwords_add
# stopwords_add로 불용어를 추가한다.

count = 0
cleaned_sentences = []
for sentence in df['reviews']:
    count += 1
    if count % 10 == 0:
        print('.', end='')
    if count % 100 == 0:
        print()

    sentence = re.sub('[^가-힣 ]', ' ', sentence)
    token = okt.pos(sentence, stem=True)

    ### 토큰화
    df_token = pd.DataFrame(token, columns=['words', 'class'])

    df_cleaned_token = df_token[(df_token['class'] == 'Noun') |
                                (df_token['class'] == 'Verb') |
                                (df_token['class'] == 'Adjective')]
    # 위 3개 클래스에 해당하는 단어만 선택.

    words = []
    for word in df_cleaned_token['words']:
        if len(word) > 1:
            if word not in stopwords:
                words.append(word)
    cleaned_sentence = ' '.join(words)
    cleaned_sentences.append(cleaned_sentence)


df['cleaned_sentences'] = cleaned_sentences
print(df.head())
df = df[['area', 'content', 'cleaned_sentences']]
df.info()
df.to_csv('./crawling_data/cleaned_merged_reviews_trip_naver.csv', index=False)
